+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,0 @@
-import { authOptions } from "@/lib/backend/auth/authOptions";
-import NextAuth from "next-auth";
-
-const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST };